requirements.txt
flask
pandas
numpy
scikit-learn
surprise
requests
python-dotenv